backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-docx==1.1.2
docxtpl==0.17.0
PyMuPDF==1.24.10
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
aiofiles==24.1.0
openai>=1.0.0
pydantic-settings>=2.13.1
google-genai>=1.68.0
jsonschema>=4.26.0
python-docx>=1.1.2
httpx>=0.27.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
alembic>=1.13.0
